Warn the user if they're specifying protocols that won't be used (#45352)
* Warn the user if they're specifying protocols that won't be used

If TLS is disabled and HTTP/1.x is available, HTTP/2 and HTTP/3 will be disabled.  This is predictable for HTTP/3, which requires TLS, but may surprise users of HTTP/2 which can be used without TLS if HTTP/1.x is not available.

Fixes #43837

* Preserve ProtocolsSetExplicitly when cloning ListenOptions

* Include endpoint in log message to reduce duplication

* Add tests

* Test LocalhostListenOptions.Clone

* Move LocalhostListenOptions into base type

...and make the fields it accesses private.  Eliminate a now-unnecessary conditional.

* Use defined logging

* Address PR feedback

Drop new linebreak.
Use the same trace level for both messages.