[Persistence] Better model for handling state persistence
Adds a `PersistentBoundary` component that allows managing the
persistence for descendant components and scopes the data to a section
of the UI to avoid conflicts between potential multiple component
instances.

Introduces the `[Persist]` attribute to let component authors indicate
what pieces of state from the component must be persisted.

Adds the ability to track removed descendant components on the renderer
to allow PersistentBoundary to efficiently remove descendant components
that registered to persist their data. This enables the components to
avoid having to implement IDisposable.

For the time being this functionality is internal to the framework and
is only used by `PersistentBoundary`

Update `ComponentBase` to check if the component implements
`IHandleComponentPersistentState` and if there is a scope available
to restore the component state if available and register for persisting
the state otherwise.