Delete Interop.FunctionalTests.ChromeTests
They've been skipped since 2018 and the linked bug was closed with the resolution that we have good coverage without doing additional work.

Why?  This is the only reason Interop.FunctionalTests pulls in Selenium, which is the only reason it can't be strong-name signed, which is the only reason product projects can't give it IVT.