In InputBase, stop mutating the Value property. Respect the one-way parameter flow like all components should.
Previously this was why setters could successfully revert inputs, but it's not the right way to do it and leads to surprising behavior (like the fact that setters could successfully revert inputs even without using SetUpdatesAttributeName). We explicitly document that components should not mutate their own inputs. Now it doesn't, and behaves more predictably but still passes tests because of SetUpdatesAttributeName.