Avoid accessing RequestServices & creating a logger in Utf8ContentHttpResult if no logging is being performed (#46560)
`Utf8ContentHttpResult` is always incurring the cost of accessing `HttpContext.RequestServices`, resolving the `ILoggerFactory`, and creating an `ILogger`, even in cases when it doesn't log anything. This has a significant cost in benchmarking scenarios, dropping RPS from ~991K to ~798K.

This change simply avoids the cost of getting an `ILogger` if no logging is to be performed.