Add LookupName to EndpointParameter. (#47272)
Separates the ```Name``` property on ```EndpointParameter``` into two properties, ```SymbolName``` and ```LookupName```. The ```SymbolName``` field is used for actual C# identifiers in code generation and is taken directly from the parameter name on the handler delegate. The ```LookupName``` field can either be the same as ```SymbolName```, or when an overriding attribute is provided (e.g. ```[FromQuery(Name="blah")]```), the overridden value is used. The name has some escaping provided to it since it's a string literal being injected into the compiled code.