Enable SingleFileAnalyzer
When publishing an app that uses WebApplication.CreateBuilder, we are getting unnecessary warnings.

We need the single file analyzer enabled in our code that is trimmable/AOT-able. When publishing for NativeAOT, the app is built into a single file, so we also need to ensure our code works correctly in single file mode.