Flow FullAssemblySigningSupported for source-build to repo tasks
This property is being added to arcade via
https://github.com/dotnet/arcade/pull/12749 and
https://github.com/dotnet/arcade/pull/12940

Once this property is added to arcade, it flows correctly to the main
aspnetcore build, but not to the build for repo tasks. The repo tasks
still need this, otherwise they end up using full signing.

Fix that by manually passing the property along (using env var) when
building the repo tasks.