Use Json TypeInfoResolverChain (#47450)
* Use Json TypeInfoResolverChain

- Removes all usages of JsonSerializerOptions.AddContext, which will be obsoleted by https://github.com/dotnet/runtime/issues/83280
- Update ProblemDetailsJsonContext to be added in a Configure<JsonOptions>() call back and to always be added to the beginning of the resolver chain at that time
    - This gives us the simplest, most understandable pattern for all libraries to follow.
- Small clean up in the API template's usings

* Change ProblemDetailsJsonOptionsSetup to use Configure instead of PostConfigure.

When adding the ProblemDetailsJsonContext, we always prepend it to the beginning of the chain at the time the configure step is executed, no matter the state of the chain.

This allows for a simpler, more understandable policy for all libraries that want to add their JsonContext into the JsonSerializerOptions resolver chain. The order of the chain is now determined by the order that the configure steps were registered in DI (for example when AddProblemDetails() was called).