Workaround for ASP.NET composite image per #84860
In our work on enabling composite Docker containers on Linux we
hit a problem caused by the fact that the SDK uses and publishes
its own version of the managed assembly

Microsoft.Extensions.Logging.Abstractions

clashing with the version of the assembly in the composite image.
As we're doing our best for publishing the Docker container in
Preview 4, I propose merging in this simple workaround as a
stopgap change; I have created a tracking issues for fixing this
cleanly by fixing SDK to start using the version of the assembly
from the aspnetcore repo instead of using its own copy.

Thanks

Tomas