Avoid delegate shuffle thunks in StringUtilities (#48234)
Avoid delegate shuffle thunks in StringUtilities

Delegates to static methods come with a performance penalty on each invocation. The code generation for delegate `Invoke` assumes the target of the delegate is an instance method. At the callsite all arguments are loaded into registers/stack that assume argument 0 is `this`. If it ends up not being the case because the method is static, the delegate `Invoke` is actually a shuffle thunk that first shuffles all the arguments and only then calls the target.

This PR stops forcing Roslyn's hand on this and replaces explicit static methods with lambdas. Roslyn knows about the static method penalty so it will generate a dummy closure class and make the lambdas instance method bodies of this dummy closure class.