[DRAFT] Consider symlinks when watching certificate file paths
Also watch resolved targets of symlinks.

There are still some rough edges around directory symlinks, since non-polling watchers tend not to notify us when they change.  I think there might be something we can do, but I have yet to investigate.

This may or may not work for the k8s case, depending on what events are produced when the `..data` symlink is updated (and whether polling is enabled).