Handle directory symlinks
Specifically, move the file watcher up one level so that we can also see changes to the parent directory of a cert, even if it's a symlink.

Note that this approach still won't handle changes to directories further up the hierarchy.  It also does nothing for files directly in the root directory which, presumably, can't be modified anyway.

The tests are failing because they're testing for things that are no longer intended to be true.