Consume IHostEnvironment.ContentRootFileProvider from CertificatePathWatcher
We shouldn't be constructing our own `PhysicalFileProvider`s in case the user is abstracting away the file system (e.g. for testing).

This simplifies a bunch of things, because there's no exactly one file provider, but we lose the ability to watch certificates outside the content root (which is probably more consistent with our design anyway).