[Blazor] Adds support for persting prerendered state on Blazor Web applications.
Persists state both for server and webassembly as necessary.
Initializes the state when a given interactive runtime is initialized
and renders the first set of components.
On WebAssembly, this is the first time the app starts.
On Server this happens every time a circuit starts.
The state is available during the first render, until the components
reach quiescence.
The approach we follow is different for server and webassembly:

On Server, we support initializing the circuit with an empty set of
descriptors and in that case, we delay initialization until the first
UpdateRootComponents call is issued.
This is because it's hard to deal with the security constraints imposed
by starting a new circuit multiple times, and its easier to handle them
within UpdateRootComponents. We might switch this approach in the future
to go through StartCircuit too.

On WebAssembly, we query for the initial set of webassembly components
when we are starting the runtime in a Blazor Web Scenario.
We do this because Blazor WebAssembly offers a programatic API to render
root components at a given location defined by their selectors, so we
need to make sure that those components can receive state at the same
time the initial set of WebAssembly components added to the page.

There are a set of tests validating different behaviors with regards to
enhanced navigation and streaming rendering, as well as making sure that
auto mode can access the state on Server and WebAssembly, and that
Server gets new state every time a circuit is opened.