Merged PR 33543: [8.0] Prevent delivery of events to disposed components
# Prevent delivery of events to disposed components

Prevents delivery of events to disposed components.

## Description

In some cases, it was possible to bypass validation rules on an EditForm by triggering a disposal and a form-submit synchronously together.

Fixes MSRC 81646

## Customer Impact

In some cases, may allow bypass of server-side validation rules.

## Regression?

- [x] Yes
- [ ] No

5.0

## Risk

- [ ] High
- [x] Medium
- [ ] Low

This change should not break any reasonable existing apps, since we don't consider it expected that events would ever be delivered to disposed components. It would never have been possible in WebAssembly - the possibility only existed in Server due to network latency, and even then only if the client is behaving very unusually. However it is theoretically possible that some app makes use of this bug intentionally, and if so they would encounter a change of behavior when upgrading to .NET 8 RC2 or GA.

Normally we wouldn't consider a bugfix like this to be "medium" risk - I'm only selecting that level because of it being introduced so close to GA.

Note that separate fixes are being delivered in patches for 6.0 and 7.0.

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A