Fix streaming into nonstreaming outlet (#50843)
# Fix streaming into nonstreaming outlet

Makes an edge case with streaming SSR and SectionOutlet work correctly

## Description

A bug was discovered when combining streaming SSR and `<SectionOutlet>` in a particular way (i.e., streaming *directly* into the `<SectionContent>`, not via any child component, when connected to a `<SectionOutlet>` that itself is nonstreaming). In this case the streamed content would not appear in the UI.

This PR makes this situation behave the same as other uses of streaming/sections, including the existing case of streaming via a child component to a nonstreaming outlet which is what our E2E tests already covered and already worked correctly.

Fixes #50804

## Customer Impact

Without this fix, developers wanting to use sections with streaming SSR would find in some cases that it doesn't work. A workaround exists (move the streaming content to a child component) but with this fix, it simply works as intended in the first place.

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [ ] Medium
- [x] Low

It's only a change to how we decide which streamed components to emit content for, so it can only affect this one functional area. The code change is just to account for a case we hadn't considered before.

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A