Remove MapIdentityApi's InfoResponse.Claims property (#51177)
## Description

This is a public API change to remove the `InfoResponse.Claims` property which was first made public in RC2. The API was approved last week during API review ([Notes](https://github.com/dotnet/aspnetcore/issues/50037#issuecomment-1740095314)).

The API was already broken in common scenarios where a user has multiple roles causing multiple claims with the same type. This cause an `ArgumentException` in the `/manage/info` endpoint when it attempted to add duplicate Key in the now-deleted `public IDictionary<string, string> Claims` property.

We had considered trying to fix the `Claims` property by making it an array or a dictionary of arrays, but @mconnew brought up the possibility that developers would not always expect clients to read their own claims. With cookies and JWE tokens claims are encrypted and potentially secret even to the authenticated user. When we consulted @blowdart, he agreed that we should avoid exposing claims from `MapIdentityApi<TUser>()` to be safe even though he hadn't personally seen an example of a claim that needed to be kept secret from an authenticated client.
 
Fixes #50037

## Customer Impact

Prior to this change, the `/manage/info` endpoint added by the new `MapIdentityApi<TUser>()` method introduced in .NET 8, would return an empty 500 response when the user making the request had a `ClaimsPrincipal` with multiple claims of the same type. This is very common when a user is in multiple roles and was [reported by a customer](https://github.com/dotnet/aspnetcore/issues/50037) shortly after the `/manage/info` endpoint was added in preview 7.  This made the entire endpoint unusable for these users.

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [ ] Medium
- [x] Low

This is simply removing new API that was first added in .NET 8 preview 7 and made public in RC 2.

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A