Remove temporary APIs now Razor compiler is updated (#51252)
# Remove temporary APIs now Razor compiler is updated

The Razor Compiler now supports `@rendermode` and `@formname` natively, so use that.

## Description

While waiting for the Razor Compiler's native support for `@rendermode` and `@formname`, we had some temporary APIs. They are less nice to use and don't perform as well. We have always planned to remove them by the time .NET 8 ships. This PR actually removes them and switches all the tests and templates over to the new native syntax.

Since this PR is changing the project template anyway, it also implements the "remove unnecessary top-level Components folder in wasm template" (#51167)

Fixes #50225
Fixes #51167

## Customer Impact

We have to remove these APIs because otherwise they end up having to be supported forever (or at least, years), and (1) they are now unnecessary and ugly, and (2) they harm perf even if you're not using them, because the way they work requires the renderer to keep checking whether you just used them.

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [x] Medium
- [ ] Low

It's a meaningful change to the syntax that also involves some template changes. However we always planned to make this change, and couldn't have done it earlier because the compiler changes were not available until now.

Also we are still waiting for a build of VS that includes the necessary change, so until then, VS will show red squigglies for the new syntax (though it still compiles and runs successfully).

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A