Fix form actions for apps deployed behind reverse proxy (#51403)
# Fix form actions for apps deployed behind reverse proxy

Makes forms work for apps deployed behind a reverse proxy.

## Description

Apps deployed behind a reverse proxy (e.g., container apps) should not try to emit absolute URLs by default because the scheme/hostname/port may differ from what is reachable from the outside world. The fix is to emit root-relative URLs.

Fixes #51380

## Customer Impact

Without this fix, apps deployed behind a reverse proxy (e.g., in ACA) would not support form posts.

## Regression?

- [ ] Yes
- [x] No

No because this only affects SSR forms, which is a new feature in .NET 8.

## Risk

- [ ] High
- [ ] Medium
- [x] Low

Low because this is only a change to how we generate the URL for a form's `action` attribute. Previously we used an absolute URL, but now we use a root-relative one. There is no other runtime change. Everything else in this PR is extra tests and updating existing tests.

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A