Update IdentityModel to v 7.0.3
We've made a fix to the IdentityModel validation logic, which has introduced a regression in .NET 8 AAD authentication scenarios. The AAD team has approved the fix and have release a new package to NuGet. This change updates our reference to the patched version, so that we avoid the regression impacting 8.0 apps.